#28 [FEAT] 잎 저장소 구매 시 로직 구현

+++ src/main/java/hackaton/ip_backend/domain/Purchase.java
@@ -34,7 +34,7 @@ public class Purchase extends BaseEntity{
 	private Product product;
 
 	@Column(name = "total_ip_amount", nullable = false)
-	private int totalIpAmount;
+	private Long totalIpAmount;
 
 	@Column(name = "total_price", nullable = false)
 	private int totalPrice;
